CViT/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
albumentations>=1.1.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.4.0
numpy>=1.21.0
